semantic change: setparam() will not ignore existing parameters
Before, if setparam() encountered an existing param, it would ignore it and NOT override it, contrary to its name. However, due to the desire to easily change a single parameter without rebuilding the params linked list, I changed the behavior of setparam() to override an existing parameter if it exists. Furthermore, I think I fixed an uncovered bug: adding a parameter to a NULL param_t.