fixed bug: mixed up order of arguments to calloc(3)
it was sheer luck I noticed this to begin with. I actually wasn't sure if that was my fault, or some wrong key combination of vim that might have mixed up the arguments. Then, I looked on github and saw it was me. It was also luck that it continued to work, because n * m == m * n. Thanks, commutative law