issue #4: declare local variables at the top of scope
This one I find it a bit difficult to apply blindly. Modern C, according to the book, doesn't use this style anymore, in which local variables must be declared at the top of functions. And for large functions, I can see why not. But there are some instances that it makes stylistic sense to me. I cannot rationalize when it does or doesn't. It's more of a gut feeling. I don't think it's a big deal, though, because it's still clean code first and foremost.