fixed obscure double-free bug
In hindsight, it makes perfect sense, but it had me stumped for a day. Essentially, after having 'convert' the CST to the AST, I would free() the CST, including the CST's pointers to the strings in the configuration file. The problem is that, instead of copying those strings for obvious reasons, those strings were just assigned to the AST. So, by free()ing the CST, I also free()d the strings in the AST, making them junk. And hence the double-free would show up when the config was reloaded.