fail hard when encountering error with kqueue
I was never quite fond of the code before, because if the kevent() call fails, it basically just loops infinitely, trashing your syslog. It seemed like a good idea at first: if it fails, let it retry itself, but I didn't bother to limit that. A limited exponential backoff algorithm might be good here, one day.