replace die() with err()
I've been going back and forth on this. At first, I had err() and I thought it was a good way to go. Then, I read some blog post and I thought CAPTURE ALL THE ERRORS! So, I created die() so that all errors went to syslog and nothing was missed. Which sounds nice, at first, but I think it has also contributed to some accidental complexity. If you're trying to reload the configuration file, you shouldn't be checking for errors in syslog. That's silly. You should check the config file for errors by running 'dyndnsd -n' and THEN if it looks good, reload the config file. That's just as effective AND simpler. And I really want to strive for simple.