fix bug in parser
The problem is basically a double-free. I was assigning the pointer to a string to more than one domain node. And upon SIGHUP, dyndnsd would free the string via the first domain node, and then subsequent frees would fail, obviously. The solution was to make sure that all nodes receive a unique copy of their strings. Then the strings in the CST can be freed as well.