remove support user/group configuration setting
because it led to the question of what happens on dynamic config reloads... change user/group too? or ignore that part of the config? neither sounded ideal, and I realized that dynamically changing user/group is niche featuritis. underprivileged user/group can still be set via compile-time option, and that's good enough for 95% of users