remove unit testing
because it doesn't seem very useful in this application. There isn't much computation to actually test, and the process of separating out functions into separate files (for the sake of testing) is making the codebase more disparate and less compact than I prefer. Perhaps that's not a very good reason, but it just seemed like a net-negative to me.