fixed a bug i introduced a while back
when I switched from autoconf to make (per jcs), I also went from bison -> yacc, which is less featureful. I could no longer pass in <ast>; it had to be a global variable. However, I forgot to adjust for that in other parts of dyndnsd, and, as well, I didn't REALLY know what I was doing. This diff makes up for that.